Button to add the first task
The new button will be shown under the Message that says "There are no tasks".
It'll create a task if pressed (specifically the firs one) and desappear.
If the list becomes empty again, the message and the button will be visible again.
It occurs because their visibility depends of the count of the count of the list items.